requirements.txt
numpy>=1.24
scipy>=1.11
matplotlib>=3.7
h5py>=3.9
k-wave-python>=0.3.5
torch>=2.1
neuraloperator>=1.0
